Scripts actualizados para importar la configuración

+++ scripts/delete-task.js
@@ -1,5 +1,5 @@
 import './components/nav-bar.js';
-import { FETCH_HEADER_WITH_AUTH, TASK_API_URL } from './modules/constants.js';
+import { FETCH_HEADER_WITH_AUTH } from './modules/constants.js';
 import { checkIfUnauthorized } from './modules/userMangement.js';
 
 checkIfUnauthorized();
@@ -10,18 +10,18 @@ const TASK_ID = URL_HASH.substring(1);
 if (!TASK_ID || isNaN(TASK_ID))
     document.location = '/';
 
-fetch(TASK_API_URL + '/task/' + TASK_ID, {
-        headers: FETCH_HEADER_WITH_AUTH(),
-})
-.then(request => request.ok ? request.json() : Promise.reject('Something went wrong deleting the task'))
-.then(task =>
-    fetch(TASK_API_URL + '/task/' + task.id, {
+ENVIRONMENT.then(async (environment) => {
+    const TASK_API_URL = environment.TASK_API_URL;
+    const request = await fetch(TASK_API_URL + '/task/' + TASK_ID, { headers: FETCH_HEADER_WITH_AUTH() });
+    if (!request.ok)
+        throw new Error ('Something went wrong deleting the task');
+    const task = await request.json();
+    const taskRequest = await fetch(TASK_API_URL + '/task/' + task.id, {
         method: "DELETE",
         headers: FETCH_HEADER_WITH_AUTH(),
     })
-    .then(() => document.location = '/')
-    .catch(error => {
+    if (!taskRequest.ok)
         console.error(error);
-    })
-);
-
+    else
+        document.location = '/';
+});
